bypass: mv implementations out of header files
