+++ client/build.sh
@@ -100,8 +100,8 @@ build_client_dynamic() {
     go build -ldflags "${goflag} -X main.BranchName=${BranchName} -X main.CommitID=${CommitID} -X 'main.BuildTime=${BuildTime}'" -o ${bin}/cfs-client-static ${dir}/sdk/sdk_fuse.go ${dir}/sdk/http_fuse.go ${dir}/sdk/http_common.go
 
     gcc ${gccflag} -std=c99 -fPIC -shared -DDYNAMIC_UPDATE -o ${bin}/libcfsclient.so ${dir}/bypass/main.c ${dir}/bypass/libc_operation.c -ldl -lpthread -I ${dir}/bypass/include
-    g++ -std=c++11 ${gccflag} -fPIC -shared -DDYNAMIC_UPDATE -DCommitID=\"${CommitID}\" -o ${bin}/libcfsc.so ${dir}/bypass/client.c ${dir}/bypass/cache.c ${dir}/bypass/packet.c ${dir}/bypass/conn_pool.c ${dir}/bypass/ini.c ${dir}/bypass/libc_operation.c -ldl -lpthread -I ${dir}/bypass/include
-    go build -ldflags "${goflag} -r /usr/lib64 " -buildmode=plugin -linkshared -o ${bin}/libempty.so ${dir}/empty.go
+    g++ -std=c++11 ${gccflag} -fPIC -shared -DDYNAMIC_UPDATE -DCommitID=\"${CommitID}\" -o ${bin}/libcfsc.so ${dir}/bypass/client.cc ${dir}/bypass/cache.cc ${dir}/bypass/packet.c ${dir}/bypass/conn_pool.cc ${dir}/bypass/ini.c ${dir}/bypass/libc_operation.c ${dir}/bypass/util.cc -ldl -lpthread -I ${dir}/bypass/include
+    go build -ldflags "${goflag} -r /usr/lib64" -buildmode=plugin -linkshared -o ${bin}/libempty.so ${dir}/empty.go
 }
 
 build_client_nodynamic() {
@@ -109,7 +109,7 @@ build_client_nodynamic() {
     go build -ldflags "${goflag} -X main.BranchName=${BranchName} -X main.CommitID=${CommitID} -X 'main.BuildTime=${BuildTime}'" -o ${bin}/cfs-client ${dir}/sdk/sdk_fuse.go ${dir}/sdk/http_fuse.go ${dir}/sdk/http_common.go
 
     gcc ${gccflag} -std=c99 -fPIC -shared -o ${bin}/libcfsclient.so ${dir}/bypass/main.c ${dir}/bypass/libc_operation.c -ldl -lpthread -I ${dir}/bypass/include
-    g++ -std=c++11 ${gccflag} -DCommitID=\"${CommitID}\" -fPIC -shared -o ${bin}/libcfsc.so ${dir}/bypass/client.c ${dir}/bypass/cache.c ${dir}/bypass/packet.c ${dir}/bypass/conn_pool.c ${dir}/bypass/ini.c ${dir}/bypass/libc_operation.c -ldl -lpthread -I ${dir}/bypass/include
+    g++ -std=c++11 ${gccflag} -DCommitID=\"${CommitID}\" -fPIC -shared -o ${bin}/libcfsc.so ${dir}/bypass/client.cc ${dir}/bypass/cache.cc ${dir}/bypass/packet.c ${dir}/bypass/conn_pool.cc ${dir}/bypass/ini.c ${dir}/bypass/libc_operation.c ${dir}/bypass/util.cc -ldl -lpthread -I ${dir}/bypass/include
 }
 
 if [[ ${build_sdk} -eq 1 ]]; then
