+++ client/bypass/cache_test.cc
